Added major tick determining in timeHelpers.
